refactor: Make some Shared types/lib to be internal
