+++ packages/core/src/Components/Loading/UI/FoLoading.vue
@@ -11,7 +11,7 @@
 <script setup lang="ts">
 import type { Animation, LoadingProps } from '@/Components/Loading/Types/Loading';
 import type { ElementName }             from '@/Shared/Types/Variants';
-import { useColor, useSize }            from '@/Shared/Lib/UseElementClass';
+import { useColor, useSize }            from '@/Shared/Internal/Lib';
 import { computed }                     from 'vue';
 
 const props = withDefaults(defineProps<LoadingProps>(), {
