+++ packages/core/src/Components/Menu/UI/FoMenu.vue
@@ -7,12 +7,11 @@
 </template>
 
 <script setup lang="ts">
-import type { MenuProps }            from '@/Components/Menu';
-import type { ElementName }          from '@/Shared/Types/Variants';
-import { menuTextPropsInjectionKey } from '@/Components/Menu/Lib/InjectionKeys';
-import { useClass }                  from '@/Shared/Lib/UseClass';
-import { useOrientation, useSize }   from '@/Shared/Lib/UseElementClass';
-import { computed, provide }         from 'vue';
+import type { MenuProps }                    from '@/Components/Menu';
+import type { ElementName }                  from '@/Shared/Types/Variants';
+import { menuTextPropsInjectionKey }         from '@/Components/Menu/Lib/InjectionKeys';
+import { useClass, useOrientation, useSize } from '@/Shared/Internal/Lib';
+import { computed, provide }                 from 'vue';
 
 const props = withDefaults(defineProps<MenuProps>(), {
     hideText:      false,
