+++ packages/core/src/Components/Select/UI/FoSelect.vue
@@ -45,7 +45,7 @@ import type { ElementName, Size }       from '@/Shared/Types/Variants';
 import { FoFilledFocused }              from '@/Components/Focus/Internal';
 import { FoLabel }                      from '@/Components/Label';
 import { useLabelType }                 from '@/Components/Label/Internal';
-import { useSize }                      from '@/Shared/Lib/UseElementClass';
+import { useSize }                      from '@/Shared/Internal/Lib';
 import { computed, useId, watchEffect } from 'vue';
 
 interface Props {
