+++ packages/core/src/Shared/Internal/Lib/IsTextAllowedForShape.ts
