+++ packages/core/src/Shared/Internal/Lib/UseClass.ts
