+++ packages/core/src/Shared/Internal/Lib/index.ts
@@ -1,0 +1,3 @@
+export * from '@/Shared/Internal/Lib/IsTextAllowedForShape';
+export * from '@/Shared/Internal/Lib/UseClass';
+export * from '@/Shared/Internal/Lib/UseElementClass';
